refactor: internal update to names to be clear about function
