+++ packages/flutterfire_cli/lib/src/firebase/firebase_apple_writes.dart
